refactor: Refactored tests into several files, Miller-Rabin to "primality" folder.
As the subject says; no major changes besides refactor. This commit marks the
end of the objectives for this branch and as such it should be merged to develop
and possibly master.

The problem of having lots of Catch2-related targets remains, and also the
executable has to be "manually" put in `/build` folder. This however is more of
an "aesthetic" problem and thus doesn't prevent the merge. It can be solved in
the future.